refactor: Rename `lib/plugins/aws/package/lib/mergeIamTemplates.js`
