+++ lib/plugins/aws/package/index.js
@@ -7,7 +7,7 @@ const generateArtifactDirectoryName = require('./lib/generate-artifact-directory
 const generateCoreTemplate = require('./lib/generate-core-template');
 const saveServiceState = require('./lib/saveServiceState');
 const saveCompiledTemplate = require('./lib/saveCompiledTemplate');
-const mergeIamTemplates = require('./lib/mergeIamTemplates');
+const mergeIamTemplates = require('./lib/merge-iam-templates');
 const addExportNameForOutputs = require('./lib/add-export-name-for-outputs');
 const { log, style, progress } = require('@serverless/utils/log');
