+++ lib/plugins/aws/package/lib/merge-iam-templates.js
