+++ test/unit/lib/plugins/aws/package/lib/merge-iam-templates.test.js
